pyproject.toml
[project]
name = "mllab"
version = "0.1.0"
description = "Machine learning playground"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (==2)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "tensorflow (>=2.18.0,<3.0.0)",
    "plotly (>=5.24.1,<6.0.0)",
    "google-genai (>=1.0.0,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"
flake8 = "^7.1.1"
